Here is a large batch of updated classes and jsps.

+++ Dictionary/src/java/org/thdl/lex/component/BaseAnalyticalNote.java
@@ -11,7 +11,7 @@ abstract public class BaseAnalyticalNote extends LexComponent implements Seriali
     private java.lang.Integer parentId;
 
     /** nullable persistent field */
-    private java.lang.Short precedence;
+    private java.lang.Integer precedence;
 
     /** nullable persistent field */
     private java.lang.String analyticalNote;
@@ -20,7 +20,7 @@ abstract public class BaseAnalyticalNote extends LexComponent implements Seriali
     private org.thdl.lex.component.ILexComponent parent;
 
     /** full constructor */
-    public BaseAnalyticalNote(java.lang.Boolean deleted, List analyticalNotes, org.thdl.lex.component.Meta meta, java.lang.Integer parentId, java.lang.Short precedence, java.lang.String analyticalNote, org.thdl.lex.component.ILexComponent parent) {
+    public BaseAnalyticalNote(java.lang.Boolean deleted, List analyticalNotes, org.thdl.lex.component.Meta meta, java.lang.Integer parentId, java.lang.Integer precedence, java.lang.String analyticalNote, org.thdl.lex.component.ILexComponent parent) {
         super(deleted, analyticalNotes, meta);
         this.parentId = parentId;
         this.precedence = precedence;
@@ -45,11 +45,11 @@ abstract public class BaseAnalyticalNote extends LexComponent implements Seriali
         this.parentId = parentId;
     }
 
-    public java.lang.Short getPrecedence() {
+    public java.lang.Integer getPrecedence() {
         return this.precedence;
     }
 
-    public void setPrecedence(java.lang.Short precedence) {
+    public void setPrecedence(java.lang.Integer precedence) {
         this.precedence = precedence;
     }
