+++ Dictionary/src/java/org/thdl/lex/component/BaseSpeechRegister.java
@@ -11,16 +11,16 @@ abstract public class BaseSpeechRegister extends LexComponent implements org.thd
     private java.lang.Integer parentId;
 
     /** nullable persistent field */
-    private java.lang.Short precedence;
+    private java.lang.Integer precedence;
 
     /** persistent field */
-    private java.lang.Short register;
+    private java.lang.Integer register;
 
     /** nullable persistent field */
     private org.thdl.lex.component.ILexComponent parent;
 
     /** full constructor */
-    public BaseSpeechRegister(java.lang.Boolean deleted, List analyticalNotes, org.thdl.lex.component.Meta meta, java.lang.Integer parentId, java.lang.Short precedence, java.lang.Short register, org.thdl.lex.component.ILexComponent parent) {
+    public BaseSpeechRegister(java.lang.Boolean deleted, List analyticalNotes, org.thdl.lex.component.Meta meta, java.lang.Integer parentId, java.lang.Integer precedence, java.lang.Integer register, org.thdl.lex.component.ILexComponent parent) {
         super(deleted, analyticalNotes, meta);
         this.parentId = parentId;
         this.precedence = precedence;
@@ -33,7 +33,7 @@ abstract public class BaseSpeechRegister extends LexComponent implements org.thd
     }
 
     /** minimal constructor */
-    public BaseSpeechRegister(java.lang.Boolean deleted, List analyticalNotes, org.thdl.lex.component.Meta meta, java.lang.Short register) {
+    public BaseSpeechRegister(java.lang.Boolean deleted, List analyticalNotes, org.thdl.lex.component.Meta meta, java.lang.Integer register) {
       super(deleted, analyticalNotes, meta);
         this.register = register;
     }
@@ -46,19 +46,19 @@ abstract public class BaseSpeechRegister extends LexComponent implements org.thd
         this.parentId = parentId;
     }
 
-    public java.lang.Short getPrecedence() {
+    public java.lang.Integer getPrecedence() {
         return this.precedence;
     }
 
-    public void setPrecedence(java.lang.Short precedence) {
+    public void setPrecedence(java.lang.Integer precedence) {
         this.precedence = precedence;
     }
 
-    public java.lang.Short getRegister() {
+    public java.lang.Integer getRegister() {
         return this.register;
     }
 
-    public void setRegister(java.lang.Short register) {
+    public void setRegister(java.lang.Integer register) {
         this.register = register;
     }
